split() and isSplit() implementation
implement split() which takes the hand's second card, removes it from internal Card list and returns it, so it can be used to create another hand
class now has a split flag which is set to true once split() has been called; getter function isSplit() added for this value. this flag is important for making sure split hands can't be counted as naturals
added new constructor that takes a Card as well as a bet value as parameters, to be used in conjunction with split() for creation of the additional hand. this constructor defaults the split flag to true