refactoring and split() functionality
all UI element instantiation is now done in a single method, initialiseUI()
GameActivity now implements the OnClickListener interface so that buttons can use GameActivity as their OnClickListener, this means all button functionality can be set in a single method, onClick()
moved the initial game state set up from onCreate to a new method initialiseGameState().
onCreate() now calls initialiseUI() and initialiseGameState()

new method split()
takes the second card of the current hand and creates a new hand with the same bet, adds the new hand to the list playerHands. draws a new card for each of the two hands so you have now have two complete hands. currentPlayerHand pointer is set to the newer hand.

consolidated win, push, and lose methods into a single new method resolveHand()
called by evaluateResult(), it pays out winnings of a hand and sets the currentPlayerHand to the next Hand in line. if there's no Hands left, calls finishGame()

new method nextHand() and helper function getNextHand()
nextHand() is called whenever we're done with the current hand. since split adds the new hand to the end of the list of hands and sets it as the current hand, when we're done with it we'll go backwards through our list for the next one. getNextHand recursively searches backwards for the next non-null item in the hands list and returns it, if it reaches the beginning of the list it returns null.

dealerTurn()
functionality used to be part of stay(), now is a seperate method. it is called when all hands have either busted or stayed and resolves the dealer hand; flips the down card and continues to hit until it reaches 17 or busts. after it's finished, calls evaluateAllResults()

evaluateAllResults()
calls evaluateResult() on each hand in the list of player hands. when it's done, calls finishGame()

evaluateResult()
evaluateResult2() (forgot to rename) does the same thing as before but now takes a Hand as a parameter. decides the outcome of a given Hand once the dealer has taken their turn

stay() functionality spread out,
stay2() (forgot to rename) now just sets the active player hand to the next one by calling nextHand, if there's no more hands then it calls new dealerTurn()