hand stayed flag
hand has a stayed flag to keep track of whether it has stayed or not. stay() to set stayed to true, stayed() to return the flag value.
used by getHand methods to prevent returning a hand that has already stayed in the case where we want only active (unfinished) hands returned