Card image display prototype
added an extra ConstraintLayout and button to the ui for testing purposes

method addCardToDisplay, pass it a ConstraintLayout(probably will be abstracted later) and a Card
it will display the given card's image in the given constraintlayout.
pressing the test button will add an ace of hearts to the display.
newer cards are placed over the top of the previous card, shifted up and right such that the rank of the card beneath is visible
upon adding a new card, adjusts all card image sizes (maintaining aspect ratio) and shift amounts such that all cards are displayed without exceeding the bounds of the constraintlayout, while ensuring all card ranks remain visible