Card Image Displays
GameActivity.java:
now using actionbar-based back button.

hands are now displayed using stacks of cards.

createCardView takes a card and returns an ImageView with the correct resource image and attribute configurations (scaling behaviour etc)
addCardToHandDisplay finds the correct ConstraintLayout for given a hand and calls addCardToDisplay to get the created ImageView and add it as a child of the aforementioned ConstraintLayout

correct display relies upon knowing the size of the play area, found by setTotalHandDisplayDimensions and setTotalDealerDisplayDimensions. these methods use a runnable in order to ensure that the dimensions are measured after the corresponding views are drawn.

calculateIdealWidth and calculateIdealHeight use the dimensions obtained by the aforementioned methods to calculate how big a card should be, and how far it should be shifted upwards and to the right of the card below it (if any) to ensure that the rank of every card is always visible.

ideally, cards take up the entire width of the container. that means that the size of the cards depends on the size of the container.
                it seems that for the current card images, we want to show about 18% of the card beneath, to see it's rank.
        this means marginSize = 0.18 * cardWidth

        the ideal card size is found by
            idealWidth = parentWidth - numberOfCards * marginSize

        since marginSize = marginProportion * idealWidth, we get
            idealWidth = parentWidth - numberOfCards * (0.18 * idealWidth)

        solving for idealWidth gives us
            idealWidth = parentSize/((0.18 * numberOfCards) + 1)

cards are shown in a ConstraintLayout so the margins of a card can be set relative to the one beneath it.
resizeCards gets the ideal size and applies it to each view in the ConstraintLayout.
setCardConstraints applies the margins to each view in the ConstraintLayout.

calibrateMagins and calibrateSizes are for re-calculating and re-applying the correct sizes and margins, for whenever things change.

bunch of other methods to make accomplishing the main tasks easier such as popCard which removes the topmost card from a ConstraintLayout, emptyCardDisplay and emptyView which clear out a ConstraintLayout, redrawHand

updatePlayerInformation removed since the method of displaying cards is now different, the task of showing and highlighting relevant hands is now covered by showActiveHandGroups and focusActiveHand

activity_game.xml:
bets are again displayed below their associated hand, as well as the net gain/loss after a round is over
flattened layout somewhat
now using guidelines to correctly portion out the screen area
utilising animateLayoutChanges on the hand displays for a more polished look